Trait zephyrus::parse::Parse

pub trait Parse<T: Send + Sync + 'static>: Sized {
    fn parse<'life0, 'life1, 'life2, 'async_trait>(
        _http_client: &'life0 Client, 
        _data: &'life1 T, 
        _value: Option<&'life2 CommandOptionValue>
    ) -> Pin<Box<dyn Future<Output = Result<Self, SlashParseError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn option_type() -> CommandOptionType;

    fn named_parse<'life0, 'life1, 'life2, 'async_trait>(
        name: &'static str, 
        http_client: &'life0 Client, 
        data: &'life1 T, 
        iterator: &'life2 mut DataIterator<'_>
    ) -> Pin<Box<dyn Future<Output = Result<Self, SlashParseError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

    fn add_choices() -> Box<dyn Fn() -> Option<Vec<CommandOptionChoice>> + Send> { ... }

    fn is_required() -> bool { ... }
}

The core trait of this framework, it is used to parse all command arguments

Required methods

fn parse<'life0, 'life1, 'life2, 'async_trait>(
_http_client: &'life0 Client,
_data: &'life1 T,
_value: Option<&'life2 CommandOptionValue>
) -> Pin<Box<dyn Future<Output = Result<Self, SlashParseError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Parses the option into the argument.

fn option_type() -> CommandOptionType

Returns the option type this argument has.

Provided methods

fn named_parse<'life0, 'life1, 'life2, 'async_trait>(
name: &'static str,
http_client: &'life0 Client,
data: &'life1 T,
iterator: &'life2 mut DataIterator<'_>
) -> Pin<Box<dyn Future<Output = Result<Self, SlashParseError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Send + 'async_trait,

Parses an argument by the option name, this is the entry point of this trait and should never be overridden.

fn add_choices() -> Box<dyn Fn() -> Option<Vec<CommandOptionChoice>> + Send>

Adds the possible choices to the argument, this function is usually implemented by the derive macro, but can be overridden manually.

fn is_required() -> bool

Sets if the argument is required, by default is true.