Module zephyrus::twilight_exports

Structs

ApplicationCommand

Data present in an Interaction of type ApplicationCommand.

BaseCommandOptionData

Data supplied to a CommandOption of type Boolean, User, Channel, Role, or Mentionable.

CallbackData

Optional extra data sent when responding to an Interaction of type ApplicationCommand.

ChannelCommandOptionData

Data supplied to a CommandOption of type Channel.

ChannelId

Unique ID of a channel.

ChoiceCommandOptionData

Data supplied to a CommandOption of type String.

Client

Twilight’s http client.

CommandData

Data received when an ApplicationCommand interaction is executed.

CommandDataOption

Data received when a user fills in a command option.

GenericId

Unique ID of a generic item.

GuildId

Unique ID of a guild.

InteractionCreate

Message

MessageComponentInteraction

Information present in an Interaction::MessageComponent.

MessageId

Unique ID of a message.

NumberCommandOptionData

Data supplied to a CommandOption of type Integer or Number.

OptionsCommandOptionData

Data supplied to a CommandOption of type SubCommand or SubCommandGroup.

RoleId

Unique ID of a role.

TwilightCommand

Data sent to discord to create a command.

UpdateOriginalResponse

Update the original response created by a interaction.

UserId

Unique ID of an user.

Enums

CommandOption

Option for a Command.

CommandOptionChoice

Specifies an option that a user must choose from in a dropdown.

CommandOptionType

Type of a CommandOption.

CommandOptionValue

Value of a CommandDataOption.

Interaction

Payload received when a user executes an interaction.

InteractionResponse

Payload used for responding to an interaction.